code cleanup continued with eslint and ts
